backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "dishka>=1.8.0",
    "fastapi>=0.129.0",
    "pydantic-settings>=2.13.0",
    "uuid>=1.30",
    "uvicorn[standard,standart]>=0.41.0",
]
